Midway implementing persist schema visualizer node positions
