+++ apps/studio/components/interfaces/Database/Schemas/SchemaGraph.tsx
@@ -1,208 +1,26 @@
-import dagre from '@dagrejs/dagre'
-import type { PostgresTable } from '@supabase/postgres-meta'
-import { uniqBy } from 'lodash'
+import type { PostgresSchema, PostgresTable } from '@supabase/postgres-meta'
+import { Loader2 } from 'lucide-react'
 import { useTheme } from 'next-themes'
-import { useEffect, useMemo } from 'react'
-import ReactFlow, {
-  Background,
-  BackgroundVariant,
-  Edge,
-  MiniMap,
-  Node,
-  Position,
-  ReactFlowProvider,
-  useReactFlow,
-} from 'reactflow'
+import { useEffect, useMemo, useState } from 'react'
+import ReactFlow, { Background, BackgroundVariant, MiniMap, Node, useReactFlow } from 'reactflow'
 import 'reactflow/dist/style.css'
 
 import { useProjectContext } from 'components/layouts/ProjectLayout/ProjectContext'
+import AlertError from 'components/ui/AlertError'
+import SchemaSelector from 'components/ui/SchemaSelector'
+import { useSchemasQuery } from 'data/database/schemas-query'
 import { useTablesQuery } from 'data/tables/tables-query'
-import { Loader } from 'lucide-react'
-import { TABLE_NODE_ROW_HEIGHT, TABLE_NODE_WIDTH, TableNode } from 'ui-patterns/SchemaTableNode'
-import SchemaGraphLegend from './SchemaGraphLegend'
+import { TableNode } from 'ui-patterns/SchemaTableNode'
+import { SchemaGraphLegend } from './SchemaGraphLegend'
+import { getGraphDataFromTables, getLayoutedElementsViaDagre } from './Schemas.utils'
+import { Button } from 'ui'
 
-type TableNodeData = {
-  name: string
-  isForeign: boolean
-  columns: {
-    id: string
-    isPrimary: boolean
-    isNullable: boolean
-    isUnique: boolean
-    isUpdateable: boolean
-    isIdentity: boolean
-    name: string
-    format: string
-  }[]
-}
+// [Joshen] Persisting logic: Only save positions to local storage WHEN a node is moved OR when explicitly clicked to reset layout
 
-async function getGraphDataFromTables(tables: PostgresTable[]): Promise<{
-  nodes: Node<TableNodeData>[]
-  edges: Edge[]
-}> {
-  if (!tables.length) {
-    return { nodes: [], edges: [] }
-  }
-
-  const nodes = tables.map((table) => {
-    const columns = (table.columns || []).map((column) => {
-      return {
-        id: column.id,
-        isPrimary: table.primary_keys.some((pk) => pk.name === column.name),
-        name: column.name,
-        format: column.format,
-        isNullable: column.is_nullable,
-        isUnique: column.is_unique,
-        isUpdateable: column.is_updatable,
-        isIdentity: column.is_identity,
-      }
-    })
-
-    return {
-      id: `${table.id}`,
-      type: 'table',
-      data: {
-        name: table.name,
-        isForeign: false,
-        columns,
-      },
-      position: { x: 0, y: 0 },
-    }
-  })
-
-  const edges: Edge[] = []
-  const currentSchema = tables[0].schema
-  const uniqueRelationships = uniqBy(
-    tables.flatMap((t) => t.relationships),
-    'id'
-  )
-
-  for (const rel of uniqueRelationships) {
-    // TODO: Support [external->this] relationship?
-    if (rel.source_schema !== currentSchema) {
-      continue
-    }
-
-    // Create additional [this->foreign] node that we can point to on the graph.
-    if (rel.target_table_schema !== currentSchema) {
-      nodes.push({
-        id: rel.constraint_name,
-        type: 'table',
-        data: {
-          name: `${rel.target_table_schema}.${rel.target_table_name}.${rel.target_column_name}`,
-          isForeign: true,
-          columns: [],
-        },
-        position: { x: 0, y: 0 },
-      })
-
-      const [source, sourceHandle] = findTablesHandleIds(
-        tables,
-        rel.source_table_name,
-        rel.source_column_name
-      )
-
-      if (source) {
-        edges.push({
-          id: String(rel.id),
-          source,
-          sourceHandle,
-          target: rel.constraint_name,
-          targetHandle: rel.constraint_name,
-        })
-      }
-
-      continue
-    }
-
-    const [source, sourceHandle] = findTablesHandleIds(
-      tables,
-      rel.source_table_name,
-      rel.source_column_name
-    )
-    const [target, targetHandle] = findTablesHandleIds(
-      tables,
-      rel.target_table_name,
-      rel.target_column_name
-    )
-
-    // We do not support [external->this] flow currently.
-    if (source && target) {
-      edges.push({
-        id: String(rel.id),
-        source,
-        sourceHandle,
-        target,
-        targetHandle,
-      })
-    }
-  }
-
-  return getLayoutedElements(nodes, edges)
-}
-
-function findTablesHandleIds(
-  tables: PostgresTable[],
-  table_name: string,
-  column_name: string
-): [string?, string?] {
-  for (const table of tables) {
-    if (table_name !== table.name) continue
-
-    for (const column of table.columns || []) {
-      if (column_name !== column.name) continue
-
-      return [String(table.id), column.id]
-    }
-  }
-
-  return []
-}
-
-const getLayoutedElements = (nodes: Node[], edges: Edge[]) => {
-  const dagreGraph = new dagre.graphlib.Graph()
-  dagreGraph.setDefaultEdgeLabel(() => ({}))
-  dagreGraph.setGraph({
-    rankdir: 'LR',
-    align: 'UR',
-    nodesep: 25,
-    ranksep: 50,
-  })
-
-  nodes.forEach((node) => {
-    dagreGraph.setNode(node.id, {
-      width: TABLE_NODE_WIDTH / 2,
-      height: (TABLE_NODE_ROW_HEIGHT / 2) * (node.data.columns.length + 1), // columns + header
-    })
-  })
-
-  edges.forEach((edge) => {
-    dagreGraph.setEdge(edge.source, edge.target)
-  })
-
-  dagre.layout(dagreGraph)
-
-  nodes.forEach((node) => {
-    const nodeWithPosition = dagreGraph.node(node.id)
-    node.targetPosition = Position.Left
-    node.sourcePosition = Position.Right
-    // We are shifting the dagre node position (anchor=center center) to the top left
-    // so it matches the React Flow node anchor point (top left).
-    node.position = {
-      x: nodeWithPosition.x - nodeWithPosition.width / 2,
-      y: nodeWithPosition.y - nodeWithPosition.height / 2,
-    }
-
-    return node
-  })
-
-  return { nodes, edges }
-}
-
-const TablesGraph = ({ tables }: { tables: PostgresTable[] }) => {
+export const SchemaGraph = () => {
   const { resolvedTheme } = useTheme()
-  const backgroundPatternColor = resolvedTheme?.includes('dark') ? '#2e2e2e' : '#e6e8eb'
-  const edgeStrokeColor = resolvedTheme?.includes('dark') ? '#ededed' : '#111318'
+  const { project } = useProjectContext()
+  const [selectedSchema, setSelectedSchema] = useState<string>('public')
 
   const miniMapNodeColor = '#111318'
   const miniMapMaskColor = resolvedTheme?.includes('dark')
@@ -217,92 +35,131 @@ const TablesGraph = ({ tables }: { tables: PostgresTable[] }) => {
     []
   )
 
-  useEffect(() => {
-    getGraphDataFromTables(tables).then(({ nodes, edges }) => {
-      reactFlowInstance.setNodes(nodes)
-      reactFlowInstance.setEdges(edges)
-      setTimeout(() => reactFlowInstance.fitView({})) // it needs to happen during next event tick
-    })
-  }, [tables, resolvedTheme])
+  const {
+    data: schemas,
+    error: errorSchemas,
+    isSuccess: isSuccessSchemas,
+    isLoading: isLoadingSchemas,
+    isError: isErrorSchemas,
+  } = useSchemasQuery({
+    projectRef: project?.ref,
+    connectionString: project?.connectionString,
+  })
 
-  return (
-    <>
-      <div className="w-full h-full">
-        <ReactFlow
-          defaultNodes={[]}
-          defaultEdges={[]}
-          defaultEdgeOptions={{
-            type: 'smoothstep',
-            animated: true,
-            deletable: false,
-            style: {
-              stroke: 'hsl(var(--border-stronger))',
-              strokeWidth: 0.5,
-            },
-          }}
-          nodeTypes={nodeTypes}
-          fitView
-          minZoom={0.8}
-          maxZoom={1.8}
-          proOptions={{ hideAttribution: true }}
-        >
-          <Background
-            gap={16}
-            className="[&>*]:stroke-foreground-muted opacity-[25%]"
-            variant={BackgroundVariant.Dots}
-            color={'inherit'}
-          />
-          <MiniMap
-            pannable
-            zoomable
-            nodeColor={miniMapNodeColor}
-            maskColor={miniMapMaskColor}
-            className="border rounded-md shadow-sm"
-          />
-          <SchemaGraphLegend />
-        </ReactFlow>
-      </div>
-    </>
-  )
-}
-
-const SchemaGraph = ({ schema }: { schema: string }) => {
-  const { project } = useProjectContext()
   const {
     data: tables,
-    isLoading,
-    isError,
-    error,
+    error: errorTables,
+    isSuccess: isSuccessTables,
+    isLoading: isLoadingTables,
+    isError: isErrorTables,
   } = useTablesQuery({
     projectRef: project?.ref,
     connectionString: project?.connectionString,
-    schema,
+    schema: selectedSchema,
     includeColumns: true,
   })
 
-  if (isLoading) {
-    return (
-      <div className="flex h-full w-full items-center justify-center space-x-2">
-        <Loader className="animate-spin" size={14} />
-        <p className="text-sm text-foreground-light">Loading table...</p>
-      </div>
-    )
+  const resetLayout = () => {
+    const nodes = reactFlowInstance.getNodes()
+    const edges = reactFlowInstance.getEdges()
+
+    getLayoutedElementsViaDagre(nodes, edges)
+    reactFlowInstance.setNodes(nodes)
+    reactFlowInstance.setEdges(edges)
+    setTimeout(() => reactFlowInstance.fitView({}))
   }
 
-  if (isError) {
-    return (
-      <div className="px-6 py-4 text-foreground-light">
-        <p>Error connecting to API</p>
-        <p>{`${error?.message ?? 'Unknown error'}`}</p>
-      </div>
-    )
+  const onNodeDragStop = (node: Node) => {
+    // TODO WHEN WE ARE BACK JOSHEN
+    console.log(node)
   }
+
+  useEffect(() => {
+    if (isSuccessTables && isSuccessSchemas && tables.length > 0) {
+      const schema = schemas.find((s) => s.name === selectedSchema) as PostgresSchema
+      getGraphDataFromTables(schema, tables as PostgresTable[]).then(({ nodes, edges }) => {
+        reactFlowInstance.setNodes(nodes)
+        reactFlowInstance.setEdges(edges)
+        setTimeout(() => reactFlowInstance.fitView({})) // it needs to happen during next event tick
+      })
+    }
+  }, [isSuccessTables, isSuccessSchemas, tables, resolvedTheme])
 
   return (
-    <ReactFlowProvider>
-      <TablesGraph tables={tables} />
-    </ReactFlowProvider>
+    <>
+      <div className="flex items-center justify-between p-4 border-b border-muted">
+        {isLoadingSchemas && (
+          <div className="h-[34px] w-[260px] bg-foreground-lighter rounded shimmering-loader" />
+        )}
+
+        {isErrorSchemas && (
+          <AlertError error={errorSchemas as any} subject="Failed to retrieve schemas" />
+        )}
+
+        {isSuccessSchemas && (
+          <>
+            <SchemaSelector
+              className="w-[260px]"
+              size="small"
+              showError={false}
+              selectedSchemaName={selectedSchema}
+              onSelectSchema={setSelectedSchema}
+            />
+            <Button type="default" onClick={resetLayout}>
+              Reset layout
+            </Button>
+          </>
+        )}
+      </div>
+      {isLoadingTables && (
+        <div className="w-full h-full flex items-center justify-center gap-x-2">
+          <Loader2 className="animate-spin text-foreground-light" size={16} />
+          <p className="text-sm text-foreground-light">Loading tables</p>
+        </div>
+      )}
+      {isErrorTables && (
+        <div className="w-full h-full flex items-center justify-center px-20">
+          <AlertError subject="Failed to retrieve tables" error={errorTables} />
+        </div>
+      )}
+      {isSuccessTables && (
+        <div className="w-full h-full">
+          <ReactFlow
+            defaultNodes={[]}
+            defaultEdges={[]}
+            defaultEdgeOptions={{
+              type: 'smoothstep',
+              animated: true,
+              deletable: false,
+              style: {
+                stroke: 'hsl(var(--border-stronger))',
+                strokeWidth: 0.5,
+              },
+            }}
+            nodeTypes={nodeTypes}
+            fitView
+            minZoom={0.8}
+            maxZoom={1.8}
+            proOptions={{ hideAttribution: true }}
+            onNodeDragStop={(_, node) => onNodeDragStop(node)}
+          >
+            <Background
+              gap={16}
+              className="[&>*]:stroke-foreground-muted opacity-[25%]"
+              variant={BackgroundVariant.Dots}
+              color={'inherit'}
+            />
+            <MiniMap
+              pannable
+              zoomable
+              nodeColor={miniMapNodeColor}
+              maskColor={miniMapMaskColor}
+              className="border rounded-md shadow-sm"
+            />
+            <SchemaGraphLegend />
+          </ReactFlow>
+        </div>
+      )}
+    </>
   )
 }
-
-export default SchemaGraph
